Reusable table parameters.
