+++ facade/src/main/scala/reactST/reactTable/TableHooks.scala
@@ -20,6 +20,10 @@ object TableHooks {
   def useTableJS[D, TI <: TableInstance[D]](options: TableOptions[D], plugins: PluginHook[D]*): TI =
     js.native
 
+  // According to documentation, react-table memoizes the table state.
+  private implicit def reuseTableState[T <: TableState[_]]: Reusability[T] =
+    Reusability.byRef
+
   def useTableHook[
     D,
     TableOptsD <: UseTableOptions[D],
@@ -28,10 +32,7 @@ object TableHooks {
     ColumnObjectD <: ColumnObject[D],
     TableStateD <: TableState[D],
     Layout
-  ](implicit
-    reuseListC: Reusability[List[ColumnInterface[D]]],
-    reuseListD: Reusability[List[D]]
-  ) =
+  ] =
     CustomHook[
       TableDefWithOptions[
         D,
@@ -46,12 +47,14 @@ object TableHooks {
       .useMemoBy(_.cols)(_ => _.toJSArray)
       .useMemoBy(_.input.data)(_ => _.toJSArray)
       .buildReturning { (props, cols, rows) =>
-        Reusable.byRef(
+        val tableInstance =
           useTableJS[D, TableInstanceD](
             props.modOpts(props.tableDef.Options(cols, rows)),
             props.tableDef.plugins.toList.sorted.map(_.hook: PluginHook[D]): _*
           )
-        )
+        Reusable
+          .implicitly((cols, rows, props.modOpts, tableInstance.state))
+          .withValue(tableInstance)
       }
 
   sealed trait TableHook extends js.Object
@@ -74,106 +77,4 @@ object TableHooks {
   @JSImport("react-table", "useGridLayout")
   @js.native
   object useGridLayout extends TableHook
-
-  object HooksApiExt {
-    sealed class Primary[Ctx, Step <: HooksApi.AbstractStep](api: HooksApi.Primary[Ctx, Step]) {
-
-      final def useTable[
-        D,
-        TableOptsD <: UseTableOptions[D],
-        TableInstanceD <: TableInstance[D],
-        ColumnOptsD <: ColumnOptions[D],
-        ColumnObjectD <: ColumnObject[D],
-        TableStateD <: TableState[D],
-        Layout
-      ](
-        tableDefWithOptions: TableDefWithOptions[
-          D,
-          TableOptsD,
-          TableInstanceD,
-          ColumnOptsD,
-          ColumnObjectD,
-          TableStateD,
-          Layout
-        ]
-      )(implicit
-        step:                Step,
-        reuseListC:          Reusability[List[ColumnInterface[D]]],
-        reuseListD:          Reusability[List[D]]
-      ): step.Next[Reusable[TableInstanceD]] =
-        useTableBy(_ => tableDefWithOptions)
-
-      final def useTableBy[
-        D,
-        TableOptsD <: UseTableOptions[D],
-        TableInstanceD <: TableInstance[D],
-        ColumnOptsD <: ColumnOptions[D],
-        ColumnObjectD <: ColumnObject[D],
-        TableStateD <: TableState[D],
-        Layout
-      ](
-        tableDefWithOptions: Ctx => TableDefWithOptions[
-          D,
-          TableOptsD,
-          TableInstanceD,
-          ColumnOptsD,
-          ColumnObjectD,
-          TableStateD,
-          Layout
-        ]
-      )(implicit
-        step:                Step,
-        reuseListC:          Reusability[List[ColumnInterface[D]]],
-        reuseListD:          Reusability[List[D]]
-      ): step.Next[Reusable[TableInstanceD]] =
-        api.customBy(ctx => useTableHook(reuseListC, reuseListD)(tableDefWithOptions(ctx)))
-    }
-
-    final class Secondary[Ctx, CtxFn[_], Step <: HooksApi.SubsequentStep[Ctx, CtxFn]](
-      api: HooksApi.Secondary[Ctx, CtxFn, Step]
-    ) extends Primary[Ctx, Step](api) {
-
-      def useTableBy[
-        D,
-        TableOptsD <: UseTableOptions[D],
-        TableInstanceD <: TableInstance[D],
-        ColumnOptsD <: ColumnOptions[D],
-        ColumnObjectD <: ColumnObject[D],
-        TableStateD <: TableState[D],
-        Layout
-      ](
-        tableDefWithOptions: CtxFn[TableDefWithOptions[
-          D,
-          TableOptsD,
-          TableInstanceD,
-          ColumnOptsD,
-          ColumnObjectD,
-          TableStateD,
-          Layout
-        ]]
-      )(implicit
-        step:                Step,
-        reuseListC:          Reusability[List[ColumnInterface[D]]],
-        reuseListD:          Reusability[List[D]]
-      ): step.Next[Reusable[TableInstanceD]] =
-        super.useTableBy(step.squash(tableDefWithOptions)(_))
-
-    }
-  }
-
-  trait HooksApiExt {
-    import HooksApiExt._
-
-    implicit def hooksExtUseTable1[Ctx, Step <: HooksApi.AbstractStep](
-      api: HooksApi.Primary[Ctx, Step]
-    ): Primary[Ctx, Step] =
-      new Primary(api)
-
-    implicit def hooksExtUseTable2[Ctx, CtxFn[_], Step <: HooksApi.SubsequentStep[Ctx, CtxFn]](
-      api: HooksApi.Secondary[Ctx, CtxFn, Step]
-    ): Secondary[Ctx, CtxFn, Step] =
-      new Secondary(api)
-  }
-
-  object Implicits extends HooksApiExt
 }
